api/requirements.txt
fastapi
uvicorn
redis
pyjwt
passlib[bcrypt]
python-multipart
pydantic_settings
